Refaktoring av VedleggVelger, splittet opp i to felleskomponenter VedleggVelger og VedleggTable
