+++ src/felleskomponenter/sideDialog/sendBrev/sendBrev.less
@@ -11,10 +11,6 @@
     margin-bottom: 0.5rem;
   }
 
-  .vedlegg_label {
-    margin-top: 1rem;
-  }
-
   .hjelpetekst {
     white-space: pre-wrap;
     vertical-align: bottom;
