+++ src/felleskomponenter/vedleggTable/index.ts
@@ -1,0 +1,3 @@
+import VedleggTable from "./vedleggTable";
+
+export default VedleggTable;
