+++ src/felleskomponenter/vedleggTable/slettFritekstvedleggModal.tsx
