+++ src/felleskomponenter/vedleggvelger/vedleggVelger.less
@@ -3,36 +3,9 @@
 @import "@{frontendModuler}core/less/core";
 @import (reference) "~nav-frontend-core/less/_variabler";
 
-.slettfritekstvedlegg-modal {
-  max-width: 600px;
-  width: 70%;
-  padding: 2rem;
-
-  .typo-element {
-    font-size: 1.125rem;
-    margin-bottom: 1rem;
-  }
-
-  .typo-normal {
-    margin-bottom: 1rem;
-  }
-
-  .knapp--hoved {
-    margin-right: 1rem;
-  }
-
-  .lukknapp {
-    display: none;
-  }
-}
-
 .vedleggvelger {
-  margin: 0;
+  display: block;
   padding-top: 0.4rem;
-
-  .vedleggtable {
-    margin-bottom: 0.5rem;
-  }
 }
 
 .vedleggvelger-modal {
@@ -49,49 +22,16 @@
     padding-bottom: 1.7rem;
   }
 
-  .vedleggtable {
+  .vedleggvelger-table {
     margin-bottom: 1.7rem;
-  }
-
-  .knapp {
-    float: right;
-  }
-}
-
-.vedleggvelger,
-.vedleggvelger-modal {
-  .vedleggtable {
     border-spacing: 0 0.3rem;
     width: 100%;
 
-    .enkeltvedlegg--border-bottom {
+    .vedlegg {
       td {
         border-bottom: 0.7px solid #d8d8d8;
         padding: 0.4rem;
       }
-    }
-
-    .enkeltvedlegg {
-      background: #f8f8f8;
-
-      td {
-        border-style: solid;
-        border-color: #e6e6e6;
-        border-width: 1px 0;
-        padding: 0.2rem 0.4rem;
-        margin-bottom: 0.4rem;
-
-        &:first-child {
-          border-width: 1px 0 1px 1px;
-        }
-        &:last-child {
-          border-width: 1px 1px 1px 0;
-        }
-      }
-    }
-
-    .enkeltvedlegg--border-bottom,
-    .enkeltvedlegg {
       .skjemaelement {
         margin-bottom: 0;
       }
@@ -121,5 +61,9 @@
         }
       }
     }
+  }
+
+  .knapp {
+    float: right;
   }
 }
