+++ src/felleskomponenter/vedleggvelger/vedleggVelger.test.tsx
@@ -7,7 +7,7 @@ import * as Nav from "../../navFrontend";
 
 import VedleggVelger from "./vedleggVelger";
 import VedleggVelgerModal from "./vedleggVelgerModal";
-import VedleggTable from "./vedleggTable";
+import VedleggTable from "../vedleggTable/vedleggTable";
 
 describe("VedleggVelger", () => {
   const mockedProps = mock<ComponentProps<typeof VedleggVelger>>();
@@ -15,46 +15,6 @@ describe("VedleggVelger", () => {
 
   beforeEach(() => {
     props = instance(mockedProps);
-  });
-
-  it("markerer valgte vedlegg ved mount", () => {
-    props.dokumenter = [
-      {
-        journalpostID: "",
-        dokumentID: "",
-        tittel: "",
-        logiskeVedlegg: [],
-        avsenderEllerMottaker: "",
-        id: "ID1",
-        dato: "",
-      },
-      {
-        journalpostID: "",
-        dokumentID: "",
-        tittel: "",
-        logiskeVedlegg: [],
-        avsenderEllerMottaker: "",
-        id: "ID2",
-        dato: "",
-      },
-    ];
-    props.valgteVedlegg = [
-      {
-        journalpostID: "",
-        dokumentID: "",
-        tittel: "",
-        logiskeVedlegg: [],
-        avsenderEllerMottaker: "",
-        id: "ID2",
-        dato: "",
-      },
-    ];
-
-    const vedleggVelger = shallow(<VedleggVelger {...props} />);
-    const vedleggTable = vedleggVelger.find(VedleggTable);
-    const { valgteVedlegg } = vedleggTable.props();
-    expect(valgteVedlegg).toHaveLength(1);
-    expect(valgteVedlegg[0].id).toBe("ID2");
   });
 
   it("rendrer en knapp med tekst 'Legg til vedlegg'", () => {
