+++ src/sider/eu_eøs/stegKomponenter/vurderingArtikkel16Anmodning.js
@@ -36,6 +36,7 @@ import {
 
 import "./vurderingArtikkel16Anmodning.css";
 import * as Skjema from "../../../felleskomponenter/skjema";
+import VedleggTable from "../../../felleskomponenter/vedleggTable";
 
 const uuid = require("uuid/v4");
 
@@ -531,8 +532,12 @@ class VurderingArtikkel16Anmodning extends Component {
           </Nav.Row>
           {redigerbart && (
             <Nav.Row>
-              <Nav.Column xs="12">
-                <Nav.Typo.Element>Vedlegg til SED</Nav.Typo.Element>
+              <Nav.Column xs="6">
+                <VedleggTable
+                  valgteVedlegg={valgteVedlegg}
+                  label="Vedlegg til SED"
+                  setValgteVedlegg={setValgteVedlegg}
+                />
                 <VedleggVelger valgteVedlegg={valgteVedlegg} onChange={setValgteVedlegg} dokumenter={fysiskeDokument} />
               </Nav.Column>
             </Nav.Row>
